fix: CI issues (part 4)
