+++ platformio.ini
@@ -31,7 +31,7 @@ test_filter = test_stm32cube
 [_env:stm32-cube]
 extends = _env:stm32
 framework = stm32cube
-build_src_filter = ${_env:esp32.build_src_filter} +<../examples/example_blink_stm32.c>
+build_src_filter = ${_env:stm32.build_src_filter} +<../examples/example_blink_stm32.c>
 
 [env:az-delivery-devkit-v4-release-arduino]
 extends = _env:esp32-arduino
@@ -107,5 +107,6 @@ lib_archive = no
 monitor_speed = 115200
 debug_build_flags = -O0 -g -ggdb
 build_type = test
-build_src_filter = -<src/> +<test/unity_config.c> +<test/test_stm32cube/>
+build_src_filter = -<src/> +<test/>
 test_build_src = yes
+lib_deps = Unity
