--- test/test_arduino/test_runner_arduino.cpp
@@ -1,21 +1,0 @@
-/**
- * @file test_runner.cpp
- * @author jake-is-ESD-protected
- * @date 2024-03-07
- * @brief Main test runner file.
- *
- * Not yet implemented.
- */
-
-
-#include "common_test.h"
-
-
-void setup(){
-    test_all(__null);
-}
-
-
-void loop(){
-
-}
